requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-on-heroku==1.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dotenv==0.19.2
pytz==2021.3
reportlab==3.6.6
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.5
